+++ authlib/integrations/flask_client/apps.py
@@ -100,16 +100,23 @@ class FlaskOAuth2App(FlaskAppMixin, OAuth2Mixin, OpenIDMixin, BaseApp):
                 "state": request.form.get("state"),
             }
 
-        claims_options = kwargs.pop("claims_options", None)
         state_data = self.framework.get_state_data(session, params.get("state"))
         self.framework.clear_state_data(session, params.get("state"))
         params = self._format_state_params(state_data, params)
+
+        claims_options = kwargs.pop("claims_options", None)
+        claims_cls = kwargs.pop("claims_cls", None)
+        leeway = kwargs.pop("leeway", 120)
         token = self.fetch_access_token(**params, **kwargs)
         self.token = token
 
         if "id_token" in token and "nonce" in state_data:
             userinfo = self.parse_id_token(
-                token, nonce=state_data["nonce"], claims_options=claims_options
+                token,
+                nonce=state_data["nonce"],
+                claims_options=claims_options,
+                claims_cls=claims_cls,
+                leeway=leeway,
             )
             token["userinfo"] = userinfo
         return token