+++ authlib/oidc/core/grants/implicit.py
@@ -80,7 +80,7 @@ class OpenIDImplicitGrant(ImplicitGrant):
     def validate_authorization_request(self):
         if not is_openid_scope(self.request.scope):
             raise InvalidScopeError(
-                'Missing "openid" scope',
+                "Missing 'openid' scope",
                 redirect_uri=self.request.redirect_uri,
                 redirect_fragment=True,
             )