+++ docs/changelog.rst
@@ -6,12 +6,16 @@ Changelog
 
 Here you can see the full list of changes between each Authlib release.
 
+
 Version 1.5.2
 -------------
 
 **Unreleased**
 
 - Forbid fragments in ``redirect_uris``. :issue:`714`
+- Fix invalid characters in ``error_description``. :issue:`720`
+- Add ``claims_cls``` parameter for client's ``parse_id_token`` method. :issue:`725`
+
 
 Version 1.5.1
 -------------